feat(weave): Add support for OpenAI Image API
